backend/requirements/common.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.17.0
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
dnspython==2.3.0
email-validator==1.3.1
exceptiongroup==1.1.0
fastapi==0.92.0
fastapi-users==10.4.0
fastapi-users-db-tortoise==2.0.0
frozenlist==1.3.3
h11==0.14.0
idna==3.4
iniconfig==2.0.0
iso8601==0.1.16
magic-filter==1.0.9
makefun==1.15.1
multidict==6.0.4
packaging==23.0
passlib==1.7.4
pluggy==1.0.0
pycparser==2.21
pydantic==1.10.5
PyJWT==2.6.0
pypika-tortoise==0.1.6
pytest==7.2.1
pytest-asyncio==0.20.3
python-multipart==0.0.5
pytz==2022.7.1
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
tomli==2.0.1
tortoise-orm==0.18.1
typing_extensions==4.5.0
uvicorn==0.20.0
yarl==1.8.2
